+++ src/components/TagCloud.css
@@ -37,7 +37,7 @@ input {
   top: 0;
   bottom: 0;
   margin: auto;
-  background: url("./cancelBlack.svg") no-repeat;
+  background: url("../img/cancelBlack.svg") no-repeat;
   background-position: center;
   cursor: pointer;
 }
